feat: Updated sample project (#368)
