+++ samples/unity-of-bugs/Assets/Scenes/1_Bugfarm.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 2cda990e2423bbf4892e6590ba056729
+guid: ea1c53ffa9f3c482287ee84d42ea0db4
 DefaultImporter:
   externalObjects: {}
   userData:
