+++ samples/unity-of-bugs/Assets/Scenes/2_MobileNativeSupport.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 8920455e39181fe469e9ddc3cef376a9
+guid: b2239cfd14b1c400ea7357ae542078c1
 DefaultImporter:
   externalObjects: {}
   userData:
