+++ samples/unity-of-bugs/Assets/Scenes/3_Transition.unity.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: e1f7138e265169147b8437fb0820f447
+guid: 9770f231579c0402c9384221ae02c07f
 DefaultImporter:
   externalObjects: {}
   userData:
