+++ samples/unity-of-bugs/Assets/Scripts/BugFarmButtons.cs
@@ -1,0 +1,88 @@
+using System;
+using System.Runtime.CompilerServices;
+using System.Threading.Tasks;
+#if ENABLE_IL2CPP || PLATFORM_IOS
+using System.Runtime.InteropServices;
+#endif
+using Sentry;
+using UnityEngine;
+using UnityEngine.Assertions;
+using UnityEngine.SceneManagement;
+
+public class BugFarmButtons : MonoBehaviour
+{
+    private void Start()
+    {
+        Debug.Log("Sample Start 🦋");
+        Debug.LogWarning("Here come the bugs 🐞🦋🐛🐜🕷!");
+    }
+
+    public void AssertFalse() => Assert.AreEqual(true, false);
+
+    [MethodImpl(MethodImplOptions.NoInlining)]
+    public void ThrowNull() => throw null;
+
+    public void ThrowExceptionAndCatch()
+    {
+        Debug.Log("Throwing an instance of 🐛 CustomException!");
+
+        try
+        {
+            throw new CustomException("Custom bugs 🐛🐛🐛🐛.");
+        }
+        catch (Exception e)
+        {
+            Debug.LogException(e);
+        }
+    }
+
+    public void ThrowNullAndCatch()
+    {
+        Debug.Log("Throwing 'null' and catching 🐜🐜🐜 it!");
+
+        try
+        {
+            ThrowNull();
+        }
+        catch (Exception e)
+        {
+            Debug.LogException(e);
+        }
+    }
+
+    public void CaptureMessage() => SentrySdk.CaptureMessage("🕷️🕷️🕷️ Spider message 🕷️🕷️🕷️🕷️");
+
+    public void SetUser()
+    {
+        SentrySdk.ConfigureScope(s =>
+        {
+            s.User = new User
+            {
+                Email = "ant@farm.bug",
+                Username = "ant",
+                Id = "ant-id"
+            };
+        });
+        Debug.Log("User set: ant");
+    }
+
+    public void BackgroundBreadcrumb() =>
+        Task.Run(() => SentrySdk.AddBreadcrumb("Breadcrumb from the background", "background task"));
+
+    public void LoadNativeSupportScene() => SceneManager.LoadScene("2_MobileNativeSupport");
+    public void LoadTransitionScene() => SceneManager.LoadScene("3_Transition");
+
+    [MethodImpl(MethodImplOptions.NoInlining)]
+    private void StackTraceExampleB() => throw new InvalidOperationException("Exception from A lady beetle 🐞");
+
+    // IL2CPP inlines this anyway :(
+    [MethodImpl(MethodImplOptions.NoInlining)]
+    public void StackTraceExampleA() => StackTraceExampleB();
+}
+
+public class CustomException : System.Exception
+{
+    public CustomException(string message) : base(message)
+    {
+    }
+}
