+++ samples/unity-of-bugs/Assets/Scripts/CloseGameButton.cs
@@ -1,0 +1,13 @@
+﻿using UnityEngine;
+
+public class CloseGameButton : MonoBehaviour
+{
+    public void CloseGame()
+    {
+#if UNITY_EDITOR
+        UnityEditor.EditorApplication.isPlaying = false;
+#else
+        Application.Quit();
+#endif
+    }
+}
