+++ samples/unity-of-bugs/Assets/Scripts/CloseGameButton.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 500071f954f4948cc98f9d24b2728c5e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
