+++ samples/unity-of-bugs/Assets/Scripts/MobileNativeSupport.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4513a698dd36a478aaa7578311efbcca
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
