+++ samples/unity-of-bugs/Assets/Scripts/MobileNativeSupport/CPlugin.c
