+++ samples/unity-of-bugs/Assets/Scripts/MobileNativeSupport/CPlugin.c.meta
@@ -55,7 +55,7 @@ PluginImporter:
     second:
       enabled: 1
       settings:
-        CPU: x86_64
+        CPU: AnyCPU
   - first:
       Standalone: OSXUniversal
     second:
