+++ samples/unity-of-bugs/Assets/Scripts/MobileNativeSupport/CppPlugin.cpp
