+++ samples/unity-of-bugs/Assets/Scripts/MobileNativeSupport/KotlinPlugin.kt.meta
