+++ samples/unity-of-bugs/Assets/Scripts/MobileNativeSupport/NativeSupportButtons.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 29d6d5fb3b7474e02bc83c032aa6d9c6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
