+++ samples/unity-of-bugs/Assets/Scripts/MobileNativeSupport/ObjectiveCPlugin.m.meta
