+++ samples/unity-of-bugs/Assets/Scripts/SmokeTester.cs.meta
