+++ samples/unity-of-bugs/Assets/Scripts/TransitionSceneButtons.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: ac0bbe003c101450eaac7ed81627abf1
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
