+++ samples/unity-of-bugs/ProjectSettings/EditorBuildSettings.asset
@@ -6,12 +6,12 @@ EditorBuildSettings:
   serializedVersion: 2
   m_Scenes:
   - enabled: 1
-    path: Assets/Scenes/1_BugfarmScene.unity
-    guid: 2cda990e2423bbf4892e6590ba056729
+    path: Assets/Scenes/1_Bugfarm.unity
+    guid: ea1c53ffa9f3c482287ee84d42ea0db4
   - enabled: 1
-    path: Assets/Scenes/2_TransitionScene.unity
-    guid: e1f7138e265169147b8437fb0820f447
+    path: Assets/Scenes/2_MobileNativeSupport.unity
+    guid: b2239cfd14b1c400ea7357ae542078c1
   - enabled: 1
-    path: Assets/Scenes/3_EmptyScene.unity
-    guid: 8920455e39181fe469e9ddc3cef376a9
+    path: Assets/Scenes/3_Transition.unity
+    guid: 9770f231579c0402c9384221ae02c07f
   m_configObjects: {}
