+++ samples/unity-of-bugs/UserSettings/AndroidLogcatSettings.asset
@@ -1,0 +1,95 @@
+{
+    "m_SelectedDeviceId": "f3089761",
+    "m_SelectedPackage": {
+        "deviceId": "f3089761",
+        "name": "io.sentry.samples.unityofbugs",
+        "processId": 14244,
+        "exited": false
+    },
+    "m_SelectedPriority": 0,
+    "m_KnownPackagesForSerialization": [
+        {
+            "deviceId": "f3089761",
+            "name": "io.sentry.samples.unityofbugs",
+            "processId": 24560,
+            "exited": true
+        },
+        {
+            "deviceId": "f3089761",
+            "name": "io.sentry.samples.unityofbugs",
+            "processId": 14245,
+            "exited": true
+        },
+        {
+            "deviceId": "f3089761",
+            "name": "io.sentry.samples.unityofbugs",
+            "processId": 14246,
+            "exited": true
+        },
+        {
+            "deviceId": "f3089761",
+            "name": "io.sentry.samples.unityofbugs",
+            "processId": 14242,
+            "exited": true
+        },
+        {
+            "deviceId": "f3089761",
+            "name": "io.sentry.samples.unityofbugs",
+            "processId": 14244,
+            "exited": false
+        }
+    ],
+    "m_Tags": {
+        "m_Entries": [
+            {
+                "m_Name": "Filter by all listed tags",
+                "m_Selected": false
+            },
+            {
+                "m_Name": "No Filter",
+                "m_Selected": true
+            },
+            {
+                "m_Name": "",
+                "m_Selected": false
+            },
+            {
+                "m_Name": "Tag Control...",
+                "m_Selected": false
+            },
+            {
+                "m_Name": "",
+                "m_Selected": false
+            },
+            {
+                "m_Name": "Unity",
+                "m_Selected": false
+            },
+            {
+                "m_Name": "CRASH",
+                "m_Selected": false
+            }
+        ]
+    },
+    "m_MemoryViewerState": {
+        "MemoryWindowHeight": 300.0,
+        "MemoryWindowWidth": 170.0,
+        "MemoryTypeEnabled": [
+            true,
+            true,
+            true,
+            true,
+            true,
+            true,
+            true,
+            true,
+            true,
+            true
+        ],
+        "MemoryGroup": 2,
+        "Behavior": 0
+    },
+    "m_Filter": "",
+    "m_FilterIsRegularExpression": false,
+    "m_SymbolPaths": []
+}
